leap: componentizing and improving direct navigation
