+++ pkg/grid/src/nav/search/Home.tsx
@@ -1,28 +1,6 @@
-import { debounce } from 'lodash-es';
-import React, { useCallback, useEffect } from 'react';
-import { RouteComponentProps } from 'react-router-dom';
-import { createNextPath, useNavStore } from '../Nav';
+import React from 'react';
 
-type HomeProps = RouteComponentProps;
-
-export const Home = ({ match, history }: HomeProps) => {
-  const searchInput = useNavStore((state) => state.searchInput);
-  const { push } = history;
-  const { path } = match;
-
-  const handleSearch = useCallback(
-    debounce((input: string) => {
-      push(createNextPath(path, input.trim()));
-    }, 300),
-    [path]
-  );
-
-  useEffect(() => {
-    if (searchInput) {
-      handleSearch(searchInput);
-    }
-  }, [searchInput]);
-
+export const Home = () => {
   return (
     <div className="h-full p-4 md:p-8 space-y-8 overflow-y-auto">
       <h2 className="h4 text-gray-500">Recent Apps</h2>
